pyproject.toml
[project]
name = "proj"
version = "0.1.0"
description = ""
authors = [
    {name = "Grayholm",email = "rodiongrayholm@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "fastapi-cache2 (>=0.2.2,<0.3.0)",
    "celery (>=5.5.3,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "pyright (>=1.1.405,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length=100